Create utilities file to refactor modules
